Added an evaluator in FELIX: SimpleOperation
This evaluator applies a simple unary operation to an input field, and
produces an output field. The unary operation type is a template argument,
and must be a class/struct that implements

ScalarT operator()(const ScalarT& v);

The class is templated on 4 parameters: in additon to the standard evaluation
type and traits, there is the underlying scalar type and the unary operation.
The purpose of this class is to be able to introduce simple variation to
the algorithm (e.g. optimize w.r.t. the log of a field instead of the field
itself), without having to introduce a new evaluator ad hoc.

In the hpp we define an alias for two partial specialization of this class,
corresponding to the exponential and natural logarithm functions, which are
defined in the same header.