A few minor mods to some of the STK mesh structures:
- added a method setAllPartsIO, which can be used to set all the parts
  as IO parts
- added the possibility to ignore side_maps (see note 1)
- fixed SideSetSTKMeshStruct: the modification_end() method of bulkData was
  called too soon.
- added the possiblity to add nodesets from sidesets also for TmplSTKMeshStruct

Note 1:

when loading, say, a 2D mesh, we may find that it already stores
side_maps. These can then be used to build the internal std::map's
that are used to determine the mapping the ID of 2D entities
into their corresponding ID on the 3D mesh. However, these maps may
be wrong in at least 2 scenarios:

- the 2D mesh was created as a SideSetSTKMeshStruct on the top side
  of an Extruded mesh with a given number of layers and it is now used as
  a upper side mesh for an extruded mesh with a different number of layers.
- the 2D mesh was created as a SideSetSTKMeshStruct on the top side
  of an Extruded mesh with a certain ordering (COLUMN/LAYER) and it is now
  used as a upper side mesh for an extruded mesh with a different ordering.

You should never use a mesh with the wrong maps. But if you do, you should
set "Ignore Side Maps" to true in the input file.