Shortening a bit the code for ResponseUtilities.
Instead of creating an RCP to the evaluator, registering the evaluator and
requiring the tag inside each clause of the if, we only create the evaluator.
The registration and tag requirement is done at the end, since it is the same
code for every response. This saves us a bunc of lines of code and makes the
whole class more readable.