Some minor modification in the FELIX problems
The main difference is in StokesFO, and concerns how we register/load/save
states. Instead of hard coding it for every state, we read the input parameters
and determine

- whether the state is a parameter based on the 'Distributed Parameters' sublist
- whether to load/save based on the 'Field Usage' parameter
- whether is a nodal or element state based on 'Field Type'
- whether the state is an extruded parameter

Based on the above, register the state in the state manager using the correct
layout, and we create one (or more) of the following

LoadStateField, LoadSideSetStateField, SaveStateField, SaveSideSetStateField,
GatherScalarNodalParameter, GatherScalarExtruded2DNodalParameter.

If for a particular run, a state is no longer required, we can simply remove
it from the discretization section (and the problem section), and it won't
be loaded/saved. More easily, we can just changed the 'Field Usage' to 'Unused'
in the discretiation section, and it won't be loaded/saved.