Completing initial implementation of debug option which computes and prints to screen the condition number of all or a particular Jacobian matrix in an Albany run.  To use this option, please add the following to your input file:
 <ParameterList name="Debug Output">
     <Parameter name="Compute Jacobian Condition Number" type="int" value="-1"/>
  </ParameterList>

(see examples/SteadyHeat1D/input.xml for an example).  Here are
some caveats, due to deficiencies of/in utilities in Trilinos
for computing the condition number of a matrix.

1.) The option to compute/print the Jacobian condition number works with the Albany as well as AlbanyT executable; however, Albany needs
to be build with Epetra on.  This is because the only routine I was able to find in Trilinos for
estimating condition numbers (AztecOO_ConditionNumber.cpp/h) is based on Epetra.

2.) The AztecOO::ConditionNumber class, which is used to compute the condition number of Jacobians in Albany,
should give a condition number that is of the same order as one computed in matlab, but not necessarily the same exact
condition number, due to different algorithms and norms used in MATLAB vs. AztecOO
(see this thread: https://trilinos.org/pipermail/trilinos-users/2013-April/003620.html ).

3.) On some tests I did, the new condition number utility in Albany appears to compute
in general condition numbers of the correct magnitude, with a few exceptions where the routine
computes a (nonsensical) 0 condition number.  I need to talk to solvers people about this.  The AztecOO::ConditionNumber
routine does not appear to be used by any routines that are called in Trilinos, so it may
have been written to not be super robust.

The current routine should be useful enough for most users -- it will give an estimate
of the right magnitude of the condition number, and you can see how the condition number
changes with changes in scaling, for example.  If a more robust/accurate condition number
estimate is sought, I suggest the interested user implement it in Albany::Application
in place of the computeConditionNumber routine.  It is unfortunate/surprising that
there don't appear to be more condition number estimators in Trilinos.