Adding printing of warning when computeConditionNumber routine returns non-zero status, which indicates something went wrong with the condition number computation, and so the value may not be trusted.  The status is non-zero in the instances I mentioned in my previous commit where the routine computes a 0 condition number, which is reassuring.
One can see the impact of a simple constant scaling of the Jacobian
matrix on the condition number using the new condition number
estimator in Albany on the examples/LCM/ParallelCubes test.
Currently the scaling is set to 1e9, which gives a condition number of
O(10) or O(100) depending on the domain.  If you change the scaling to 1e0
(no scaling), the condition number goes up to O(1e11).