LCM: Enabling Yaml input. (#74)
If Teuchos is compiled with Yaml support, Albany's solver factory
and LCM's material database can optionally be read as Yaml files.

This is determined by the file extension. If the extension is .xml,
the XMl parser for Teuchos ParameterList is invoked. If the extension
is .yaml, the Yaml parser is invoked.

Add small example (test) specifying the same problem with XML and Yaml
formats. The readability of the input using Yaml is significantly
improved.

Only LCM is tested for this. A cursory search finds that ATO and QCAD
call the XML parser for their own purposes. Adding support for
the Yaml parser is not difficult, following the example of the LCM
material database.

Enable Yaml support by default on Fedora.