use jac->rowMap() to allocate scaleVec_
instead of jac->colMap() in computeGlobalJacobianImplT.
presumably the two maps have the same size in serial which
is why this didn't cause any issues during our serial
testing. In parallel, however, trying the the combination
SingleStep + Diagonal scaling causes an exception to be
thrown, noting scaleVec_ does not have the same dimension
of the matrix it's trying to be applied to. using the
rowMap() fixes this issue in parallel and appears to
retain the same desired behavior in serial.