LCM: Modifying gold file to reflect new initial conditions that are inconsequential
We store the same element quantities on every element block.

For this test, element blocks 1 & 2 are bulk element blocks and
element block 3 is a surface element block.

Bulk elements have stresses denoted CAUCHY_STRESS_01, ...
Surface elements have stresses denoted SURF_CAUCHY_STRESS_01, ...

This test was failing because we changed the initial condition of variables that don't
belong on those blocks. We moved from the initial concentration to just zero.

I do like zero better and consequently, I'm updating the gold file. All the data that
matters did not diff between the gold file and the results file. In addition, we might be able
to update the truth tables in the future so that these comparisons are never made in
exodiff.