Code tweaks needed to run a coupled Albany-Peridigm problem with an applied body force.
Flipped the sign on Peridigm internal force (residual and tangent).

Hard coded a body force, currently pound-defed out.  To the best of my knowledge, it is
not possible to apply a body force to a mechanics problem in Albany.  Lack of resources
led to the hard-coding approach.  Avert your eyes.