LCM: WIP Make build scripts+modules and CDash work together [#57].
Here is what I would like to accomplish with this, but have been
unable to do so for lack of in-depth knowledge of CTest:

The files ctest.cmake and Project.xml in the Albany/doc/LCM/dashboards
directory are templates that similarly to the script build.sh are copied
or symlinked to the top level LCM directory. Not sure that Project.xml
is needed because it appears that the values set in it can also be
set within a Cmake script like ctest.cmake.

When build.sh is run in test mode as test.sh, the templates ctest.make
and Project.xml are copied into the build directory, for example
albany-build-serial-gcc-release. Then build.sh replaces actual values
for the build into the placeholder lcm_* strings in the templates.

The script build.sh then proceeds to call CTest which in turn runs
the tests and submits the results to the dashboard.