LCM: progress towards issue #11.
Adding input files to run Newmark Beta scheme through Piro and
Tempus in explicit mode (beta = 0.0, gamma = 0.5).  Studies of this scheme's
performance are still underway, but results are promising.  For
single cube geometry, my "back-of-the-envelope" calculations say
that the stable time-step is dt <= 7.211e-5; scheme is stable for dt = 7e-5,
unstable for dt = 8e-5.  For larger cube geometry, my "back-of-the-envelope" calculations
say that the stable time-step is dt <= 1.89e-5; scheme is stable for dt = 1e-5,
unstable for dt = 2e-5.

Some caveats:

- The essential BCs in Albany are not implemented in a way that will be handled correctly
with an explicit time-integration scheme at the present time (see issue #13).  This problem works explicitly
b/c there are no essential BCs in the problem specification.  Do not try explicit Newmark Beta
schemes w/ essential BCs, as this will not work correctly in general.

- The explicit Newmark Beta scheme in Piro and Tempus has not been optimized to do mass-lumping
and a trivial inversion of the resulting lumped (diagonal) mass matrix.  Since inversions of
the mass matrix are happening under the hood when running explicit, please do not expect
explicit to be faster than implicit at the present time.  The optimizations involving lumping
can be done in a straightforward way, once the scheme w/o lumping is tested more thoroughly.