LCM #53: Hijack and rework Dave Littlewood's NOX pre-post operator to handle also the Schwarz termination criterion.
It appears that there can be only one NOX pre-post operator, thus
there was a conflict between the model evaluator status flag operator
and the Schwarz termination criterion operator.

So, for now, add the functionality needed for Schwarz to the status
flag. This requires performing some operations to compute norms of
the solution at various times and adding some storage to keep these
values for later use in the Schwarz loop. Thus, this does not affect
a non-Schwarz analysis.

Still, it is probably a good idea to come up with a plan to handle
more than one NOX pre-post operator, so that this situation does
not repeat itself. And also to avoid out-of-control growth of
this class if more pre-post functionality is needed in the future.