LCM #53: First working version with convergence criterion from Schwarz paper. Issues fixed:
- Conflict between Schwarz convergence criterion and status test
NOX pre/post operators resolved with folding both into the same
class. In addition, when trying to create one or the other, check
for the existence of a pre/post operator. If one exists already,
skip the creation and use the existing one as it supports both the
status test and Schwarz convergence criterion. Still need to discuss
at large what to do in the long term to avoid out-of-control
bloat in case more functionality is needed.

- The NOX solver resets the solution to zeros before each non-linear
solve, but keeps a copy of the old solution in a previous solution
group. Use this group and the current group to compute the norms that
define the Schwarz convergence criterion. Store this info in
the pre/post operator class.

- Keep an array of the pre/post operators for each subdomain when
the solvers are created. These operators are the only way to access
the solutions and their corresponding norms in the Schwarz loop
that comes much later.