LCM #53: Add test for the Schwarz alternating implementation.
Things to do:

- Find a way to avoid exodus output at every Schwarz iteration.
This requires taking control of the output away from the nonlinear
solver and give it to the Schwarz method.

- Continuation. It appears that doing this through LOCA would be
too difficult. A suggested way to do this is through Anderson
acceleration, a variant of fixed point methods. But now that Tempus
is enabled, perhaps it would be easier to do continuation through
dynamics. Then quasi-static problems would be run as truly
quasi-static: very slow dynamic problems.

Sice we now the root cause of the unmatched convergence for the
old Schwarz implementation, I'm closing this issue and opening
new issues for the above.