LCM: Schwarz alternating:
- Rework the Schwarz loop and convergence criterion. The convergence
criterion has separate functions, much like MiniSolver. This is so
that if new or different convergence criteria are desired in the future,
they would be relatively easy to implement.

- Why the #!$* is EvalModelImpl in the ModelEvaluator a constant
method? It is natural to expect that any model will change when
evaluated. Thus, any data member that changes during EvalModelImpl
or methods called therein needs to be mutable, causing a lot of confusion.

- For some strange reason the Schwarz alternating test was failing
in some platforms even when the Schwarz loop converged fine.
It appears that it is too much to expect that the errors are the
same everywhere up to 15 significant digits, following the Foulk doctrine.
Change the test so that the solution is required to converge to
a relative error of less than 10^(-15). But don't compare the
error to a baseline and accept the solution as long as the error is
less than 10^(-15).