Work towards issue #116 (solution_dot not written to Exodus output file for transient problems). solution_dot now written to the Exodus output file for problems which:
- Are transient and hence have solution_dot.
- Are run with AlbanyT (I will add this capability to work with the Albany executable shortly).
- Use the Albany::StatelessObserverImpl class for the observer.
- Use the Albany::OrdinarySTKFieldContainer class (solution_dot will not be written to the output
file if using the Albany::MultiSTKFieldContainer; warning is printed about this in this case).

Caveats:

- This capability is primarily of interest for LCM, so getting it working for LCM problems
is my primary focus.  Spot-checking this shows that it works for basic transient problems too.  It
does not seem to work for Aeras problems, presumably b/c they may go through a different observer.
I am probably not going to make sure the writing of solution_dot works for non-LCM problems at this time,
as it is not a priority.
- I did not add this capability for problems utilizing Albany::MultiSTKFieldContainer.  It seems
no ctest tests use that class at the present time.