LCM: Addressing issue #129.  It looks like solution_dot was only created erroneously in the output file for Schwarz coupled problems.  Modifying the Schwarz::Coupled class such that x_dot is not required, which it should not be for non-transient problems.
Output of solution_dot is controlled by "Number Of Time Derivatives" in input file.
This value should be equation to 0 for non-transient problems.  Now that the 0 value
is valid for Schwarz problems, updating input files for these problems to set Number Of Time
Derivatives to 0, so as not to create x_dot in input file when it is not relevant.