Completing work towards issue #128 (populating solution_dotdot in Exodus file).
This field now gets populated for Transient
problems run with a second order time-integrator.  Not
closing this issue yet as I plan to do a bit more testing of
the new capability.

Caveats:

- Currently solution_dotdot only gets populated in the Exodus file
for Tempus problems.  It can be extended to Piro second order time-integrators
too, but would require changes in Piro.  I will look at what
this would entail on Monday and may add that capability if it is straightforward.
- Population of solution_dotdot in the Exodus file works for problems
run with the AlbanyT (not Albany) executable only.  To extend the
capability to the Albany executable, change in Trilinos Epetra packages would be required,
and I am not going there...