LCM: Rebaselined (again!) failing test. #135
The force_x value is of the order 10^2, and the force_y and force_z
values are order 10^-5. The _y and _z values diff on some platforms.
This doesn't seem like it should happen, but the values aren't that
significant... Raising the floor should fix the test failure pending
further investigation.