LCM #134: A case where the cure proved worse than the disease:
The functionality of the Schwarz convergence criterion and the NOX status
test flag were folded into a single class because NOX supports only one
PrePost operator.

Dan Ibanez pointed out that among the supported operators there is a
vector that can contain other operators, thus enabling a mechanism to
have more than one PrePost operator in NOX.

So, in order to clean up the design, the Schwarz convergence criterion
and NOX status test were split into two PrePost operators that would
be carried by the vector of PrePost operators.

Unfortunately, the functionality of this vector of PrePost operators is
extremely limited. It acts as a sink in the sense that operator objects
can be pushed into it; but once inside there is no way to know what went
in there.

This means that there is no way to verify that a Schwarz convergence
criterion object or a NOX status test flag already exists in order
to avoid creating repeated instances.

Complex logic and code would ne needed to deal with this situation,
making it worse than accepting that a single object carries
the functionality of the Schwarz convergece criterion and the NOX status
test.

So, we go back to the combined class, with a few tweaks. Also, we rename it
as SolutionSniffer because that is what it does: lies in wait, tracking
changes to the solution and acting on them.