LCM: Add -Wno-terminate to silence warnings about throwing from destructors. This is temporary until Trilinos issue 1303 is resolved.
This is not really kosher because in C++11 destructors should not
throw exceptions, but our old friend TEUCHOS_TEST_FOR_EXCEPTION insists
on doing so, among others.

As a result, when compiling, the error log gets so cluttered by
warnings about throwing from destructors that it becomes unusable.

This change is a temporary workaround to be able to use the compile
error log.

[#22]