LCM [#145]: First working version of strongly enforced Dirichlet/essential boundary conditions. It apears to work for simple mechanics problems. Creates symmetric, positive definite, and well-conditioned Jacobian. No need for preconditioners.
Needs more testing, and then writing of the corresponding
strongly-enforced DBC evaluators for the other remaining DBCs
(time dependent, torsion, etc.).

**WARNING***: The strongly enforced DBC evaluator **forcefully** changes
the solution vector by the use of a combination of aliasing and
const_cast, thus subverting the design and rules of the ModelEvaluator.

The "correct" solution would be to use something like the NOX PrePost
operator, thus requiring the coordination of this operator with the
BC evaluator. This would lead to code bloat, complication, and obfuscation.

Thus, we regard the inability to change the solution where we need as
a design flaw, and intent to keep this solution unless they point a gun
to our heads. And even then we might prefer death before dishonor.