LCM [#145]: First cut of working time-dependent SDBC.
One consequence of modifiying the solution outside of the framework
imposed by the ModelEvaluator is the following:

A LOCA continuation problem with a single step cannot possibly be
aware that the solution will be changed by the SDBC, and thus flags
the solution as converged and does nothing. This does not happen with
more than one continuation step, where the solutions are computed
correctly.

Using continuation with a single step seems like a rather contrived
situation, and because there are bigger fish to fry, we move onto
inroducing SDBCs to dynamics.

If this becomes a problem, please create a new Albany issue.