LCM: Added body force evaluator and supporting infrastructure.
There was already some infrastructure in place for the corresponding
computation of the residual, but strangely enough, there was no
LCM evaluator for body force.

I was going to model LCM's body force after ATO's version. In ATO,
the body force vector is a product of mass density and acceleration.
This is the most common case and some finite element formulations
use this decomposition as well, as it leads to some nice simplifications.

But this would complicate the use of body forces that result from
interactions other than gravity (e.g. electromagnetic).

So, for LCM, the body force is specified as a general force per
unit volume. In addition, the body force is specified per element block,
so that different materials can incorporate different interactions.

The body force is handy for debugging and verification of dynamics.

Need to add test for this soon.

[#147]