Enabled Albany-Peridigm problems with Neumann BC for simulations containing multiple elements types.
This was tricky because Albany is currently limited to a single Neumann field manager, which must
be created for a specific element topology.  Simulations containing multiple element types and Neumann
BC will almost certainly fail because the Neumann-related evaluators will assume array lengths that
coincide with the element topology for which they were created.  I will create a bug report for this.

To enable coupled Albany-Peridigm problems, which almost always have multiple element types, I added
a check such that the Neumann field manager acts only on worksets containing side sets.  This is
currently wrapped in a PERIDIGM #define.