LCM: creating version of computeGlobalResidualImplT in which initial residual is computed with the right dirichlet BCs, for the strong form of the Dirichlet BCs, to give the correct displacement with the d-form of the Newmark Beta scheme in Tempus. The call to this routine is currently commented out as it is work in progress: further testing is needed, and the routine needs to be modified to be called not just in the first residual computation, but the first residual computation of every time step. To try it out for now, user can uncomment
"this->computeGlobalResidualTempusSDBCsImplT("

(line 1607) in Albany_ApplicationT.cpp and recompile.  Caveats:

1.) The new function is WIP and logic to have it work seamlessly w/ all Albany
problems has not been put in yet.  Therefore if the code in line 1607 is uncommented,
problems not using LCM + Tempus Newmark Beta d-form + Strong DBCs probably will fail.

2.) Scaling will not work with the new function at the present time.

3.) tempus-newmark-d-form branch of Trilinos is needed for the new function but **ONLY**
for the d-form of the Newmark-Beta scheme, which currently exists only in that branch (will be
moved to master soon).  No NOX changes to Trilinos are required for the new function.

4.) New function is not designed to work with "Single Step" nonlinear solver, but rather "Line Search Based"
nonlinear solver.

This path towards getting the Newmark scheme + Strong DBCs seems promising, but work is needed
to make sure it will pan out as the solution to the problem we were experiencing.  The function has been added in
a way that it can easily be removed if the approach does not pan out.  WIP.