LCM: hooking in new routine computeGlobalResidualTempusSDBCsImplT so it gets called when running with "Newmark Implicit d-Form" Tempus stepper. This capability is now tested in the LCM/Schwarz/Dynamic/CubesSDBC test.
Still need to put in logic to force user to use d-Form of the Newmark scheme
if and only if utilizing the SDBCs (strong Dirichlet BCs).

It may be required to extend the new routine and how it is integrated into
Albany_Application.cpp for LOCA, which I will look into.