Putting in logic to throw exception when running unsupported schemes with SDBCs and standard DBCs, which required some changes to the problem classes in Albany.  Everything builds/passes on my machine.  If a compilation error is encountered regarding the "bool useSDBCs()" virtual function being pure in a prolem class, the following line nees to be added to the problem *.hpp:
virtual bool useSDBCs() const {return false; }

This should only cause problems if there is some problem class I missed due to not
building it in my nightly tests.

The new logic will correctly throw an error when running the d-form of the Newmark
scheme without SDBCs, or the a-form of the Newmark scheme with DBCs for the
Mechanics problem.  I'll propagate this logic to other LCM problems shortly.