LCM: Extend the solution sniffer to keep a copy of the solution for later use. This is handy for manual continuation in Schwarz. Modify the manual continuation accordingly, keeping an array of pointers to solutions for each subdomain.
At the first step in the continuation, the model for each
subdomain uses its default nominal values. Once a solution is
computed, it is saved in the array mentioned above.

The following steps in the continuation set the nominal values
for the solution to the saved solution from the previous step.