Added support for 3d coordinates in stk meshes
If the mesh is already 3d, then the new stored field pointer coordinates_field3d
points to the same objiect pointed by coordinates_field. However, if the
mesh dimension is lower (1d or 2d), then the 3d field points to a different
field in the metaData, which stores the coordinates as a 3d field. In general,
these are the same coordinates, padded with zeros. However, this may not
be the desired behavior in some cases. For instance, a non-flat 2d mesh
(such as the surface mesh of an extruded mesh) may need to store all the
coordinates of the points, in order to be loaded correctly the next time.

Albany::IossSTKMeshStruct can load the field "coordinates3d" if it is defined
in the mesh. This field (and NOT "coordinates") should then be used to create
the correct data structures for differential geometry.