Bugfix in SaveStateField.
Since some problems save the time as a state, we need to deal with workset
states, to avoid segfaults. This is because a workset state only has one
value per workset, so looping over the number of cells in the current
workset is wrong. On the other hand, for elem/node states, we need to loop
over that number (rather than using the first extent of the field), because
some workset may have fewer elements than the usual workset size.