Added test for side set equation, a simple laplacian on the side.
This test is comprised of 2 tests: a laplace problem in 2D on a 2D mesh, and
a laplace problem on the side of a 2D mesh. The response of the two problems
is the squared L2 norm of the solution (on the "volume" in 2D and on the side
in 3D), which is checked for correctness.