Moved DOFDivInterpolationSide to FELIX
This evaluator does not "really" compute a divergence on a sideset, meaning
that it does not use differential geometry formulas. Instead, we compute the
divergence as if we were on a planar 2D geometry (with z=0), by projecting
the intrinsic gradients in 3D and neglecting the 3rd coordinate.

Given this "peculiarity", we decided to remove it from the main PHAL folder.