LCM: partial fix to SDBC not working in parallel.
In particular, fixing the issue that the residual had nans in it.
The problem had to do with my fix in Albany::Application to make
the Jacobian/residual aligned in the first Newton step - an import
was missing for the solution workset.xT.

Modifying TimeDependentSDBC test to run larger problem in parallel
on 4 procs now that it runs and gives the correct answer.

There is still an issue for parallel SDBCs, however, namely, as I suspected,
the Jacobian rows do not get zeroed out correctly for parallel runs.
I will debug this next.