LCM: Restart NOX and pass in previous solution by using reset. This simplifies the code somewhat. Replace TEUCHOS_TEST_FOR_EXCEPTION with ALBANY_ASSERT. Rename Schwarz loop functions to reflect what they do instead of what they use. Other changes to try to simplify the LOCA Quasistatics Schwarz loop.
The NOX delay previously noted seems to exist also in non-Schwarz
runs in the master branch, thus it appears to be unrelated to Schwarz.
Not sure if this is recent behavior or it has been there all along.

Quasistatic Schwarz with manual continuation for controlling loop
and LOCA continuation for subdomains is ready for application testing.

To use it, both Trilinos and Albany need to be in the lcm-mods branch.
Any brave souls?

[#126]