FELIX: fix towards issue #187 having to do with velocities being wrong near processor boundaries for parallel runs.
The fix works on several tests I have tried on my workstation and Edison
when run with Tpetra + no reduced communicator, as that is the specific
case I was focussing on when creating the fix.

The file with the fix is separate rather than being in felix_driver.cpp for now
b/c the fix needs to be propagated into all possible configurations of CISM (Epetra, Reduced Comm builds).
I will do this soon.  For now, to use the fix, please over-write felix_driver.cpp
with felix_driver.cpp.PARALLEL_FIX, and recompile Albany and CISM, making sure
Reduced Comm and Epetra are OFF.  I have run all CISM-Albany test with the fix
and everything passes.