LCM: progress towards github issue #191.
It looks like several of my commits from 9/29-10/2 did not get
merged from the lcm-mods branch of Albany to the master.  Unfortunately,
I need to recover these and push them manually now.

This commit is the equivalent of the following to lcm-mods:

https://github.com/gahansen/Albany/commit/1cf3e8523602038fad86493511e16dcb9a7eb3b6

LCM: correcting logic in computeGlobalResidualSDBCsImplT affecting quasi-statics.

There was an error in the logic that was causing the residual to be computed
twice (to ensure proper SDBCs are applied) in every Newton solve, not just the first one after incrementing the time.  This did not mess up the solution but would have slowed down the simulation.

Some work on correcting logic in computeGlobalResidualSDBCsImplT
with Alternating Schwarz.  Making it so residual is computed
twice (to ensure proper SDBCs are applied) only in first Newton solve after
we have switched domains and in the first Newton solve after incrementing
the time, which is how I believe the algorithm should proceed.  It looks like
there may be a few extraneous residual computations for quasi-statics,
perhaps due to how the setDBCTime routine works -- I need to dig into that
to understand how this works.

For dynamics, I believe the logic described above is correct but Schwarz Alternating still does
not converge, again perhaps due to setDBCTime.  WIP.