Adding option to STK mesh to take a mesh and shift it by some fixed value in the x, y, and/or z directions.
The idea is with this capability, one can create an internally-generated
STK mesh in Albany of a domain of the form (xL, xR) x (yL, yR) x (zL, zR),
rather than (0, xR) x (0, yR) x (0, zR).

For an example of usage, please see

tests/small/LCM/DynamicsTempus/Clamped/clamped-stk-shift.yaml

Unlike in a car :), the "stk-shift" option should simplify things for Schwarz --
one can now use the STK mesh option to generate subdomains for Schwarz problems
on box geometries, which makes it easier to play with the size of the overlap region.
The new capability will be useful in creating a Schwarz version of the clamped problem.