LCM: creating SDBC + D-Form of Newmark-Beta version of Clamped test case (no Schwarz yet).
This is because we will want to use SDBC + D-Form of the Newmark-Beta scheme
for the dynamic Schwarz version of this problem, so we should test first
that it works on the regular version of the problem.

Solution computed with D-Form + SDBCs is same to machine prevision as
that computed with A-Form + DBCs.  The former one runs much slower b/c
the Nonlinear Solver is "Line Search Based" rather than "Single Step".