LCM: bug fix in Schwarz Alternating Dynamics loop.
The initial condition was not being set correctly for the case
of a non-zero initial condition - needs to be set from model evaluator
nominal values rather than from STK mesh.

With this fix, clamped problem with dynamic schwarz appears to be working,
at least for a short simulation evaluated in the "eyeball norm"!
Need to run longer in time and in parallel, as well as look at conservation
properties to verify correctness.