LCM: creating test case for dynamic Schwarz + clamped problem.
Dynamic Schwarz appears to be working - gives same solution as non-Schwarz
for the test case considered so far.

Still need to run a finer problem longer in time and verify conservation, etc.

TODO: we should come up with a way to test alternating Schwarz that does more than check
that the Schwarz iterations converged.  I worry that something might change in the code
that will alter things but not the final convergence status, and so it would not get caught by the testing.
I guess using exodiff w.r.t. a gold file would be one way to do this.