LCM: modifying DTK error function to print out error in vector solution not just each component.
I think the relative error in a particular component (e.g., z-displacement) can be misleading in the
case the norm of that component in the reference solution is close to 0.  In the case
a structure is moving all over the place, error in the norm of the displacement is probably
the most reasonable metric to look at.

Also checking in some bash scripts to post-process output from error executable, namely
extract errors for various components.