LCM: modifying how relative error written to exodus file is computed in dtk_interp_and_error utility.
Before, the pointwise value of the absolute error was divided by the pointwise value
of the reference solution.  This can cause problems in the case the reference solution
value is very small (there are dots with high error in the exodus file, which is misleading).
Changing computation to divide absolute error by norm of reference solution to produce
much smoother and more meaningful plots of absolute error.