LCM: modifying Clamped problem to be truly 1D (w/o Schwarz for now).
Also changing material parameters such that the Young's modulus E = 1e9, which was not
in some of the tests.

I will do this for the Schwarz version of this problem as well once I verify that the
correct results are obtained with today's modifications to StrongSchwarzBC (some that
have not yet been pushed).