LCM: cleaning up dynamic Schwarz updating of IC vectors and writing of solution to Exodus file.
This clean up is accomplished by writing a new function, setDynamicICVecsAndDoOutput.
It is a bit more complex than its quasi-static analog as the initial time step case
is slightly different than the subsequent steps (I wasn't able to come up with a straightforward
way to avoid this).  Regardless, doDynamicSchwarzLoop routine is much cleaner thanks
to this function.