LCM: progress towards issue #230 (implement analytic mass matrix in Albany for composite tets + dynamics).
This commit is basically taking the relevant / working code relevant to this issue
from the lcm-comp-tet-exact-mass branch (cleaned up a bit) and putting it into master.

The new framework to specify the analytic mass has been tested using an analytic mass
for hex8 elements and identical results are obtained as when using AD to get the mass
matrix.

I have not tested the new code yet with composite tet elements but will do so soon.
The framework should be correct, but there may be minor bugs (e.g., typos in the
values of the local analytic mass matrix entries).