LCM: to resolve issue #252, turning off writing of solution_dtk to Exodus output file unless we are dealing with a Schwarz problem.
It is possible to turn on writing of solution_dtk to Exodus output file
for any problem by specifying

Output DTK Field to Exodus: true

in the input file under "Discretization".

The only reason why one might want to do this is if there is a failed
comparisong using exodiff due to the gold Exodus file have solution_dtk
as a field in it.