LCM: adding new option relevant for AnalyticMassResidual: 'Lump Analytic Mass'.
The option is false by default.  It is only relevant for the AnalyticMassResidual
(when 'Use Analytic Mass' = true).  When 'Lump Analytic Mass' = true, a lumped mass will be
created in AnalyticMassResidual, rather than a consistent one.  This option
allows one to use a lumped mass for implicit problems, which is not possible with the
lumping done in Piro.  This option may be of interest for, e.g., comparisons to Sierra,
which only lets one use a lumped mass, even for implicit dynamics.