LCM: Add missing bulk modulus.
Max Bloomfield from RPI discovered that the bulk modulus was missing
from the thermal expansion term for the stress in the material
models listed below.

Add the missing bulk modulus in the stress calculation. Also, add
it to the energy for the Hyperelastic damage model.

There may be more models with this problem. I'll look for them,
but wanted to fix these right away, since they are the ones
more frequently used.