LCM: Aborted attempt shards::Array->Kokkos::View
Most of the code compiles, and the Albany executable appears
to run fine with this replacement.

Still, the shallow copy problem in Schwarz persists.
In addition, there are compile errors with some
conversions from STK.

I don't have time to go down that rabbit hole,
or to wait until I get approval to change shards
to fix the memory leak and add a deep copy.

So, I'll attempt to copy the shards:Array files to Albany
and add a way to do a deep copy.