LCM: Copy internal states, third try.
Abandoned idea of adding a deep copy operator in shards.
It wasn't going to work, either.

Essentially, the raw pointer in shards:Array in intialized to
a contiguous area of memory that STK holds. This memory contains
the values of the states, old and new.

When old and new states are swapped, all happens within that STK
storage space: nothing is deleted or allocated.

Thus, in order to create a copy of the internal states that can
be later used to overwrite Albany::StateArrays, one needs to
allocate separate space for this, copy the entries of all state
arrays and entries one by one. When the need to overwrite arises,
these saved values are used to overwrite the Albany::StateArray
values, which in turn are stored in STK because of the raw
pointers mentioned above.

To try to make this easy and clear, there are now counterparts
to Albany::StateArrays called LCM::StateArrays. The latter have
actual storage in the form of std::vectors.

There are also a few helper functions that copy to and from these
arrays. When copying from Albany to LCM, the LCM arrays will be
either created or resized to contain whatever the Albany arrays
contain, and then the copy is performed. When copying from
LCM to Albany, it is verified that the LCM arrays are of the same
size and contain the same number and type of internal variable
strings before copying anything, to ensure that nothing is
being corrupted.

With this, the cuboids problem finally produces the same results
in single domain and two domains in Schwarz for J2 plasticity:
the values of eqps are the same to numerical precision.

More testing is needed to verify that this works generally
for inelasticity.